+++ src/runtime/src/core/io.rs
@@ -95,7 +95,19 @@ pub fn make_module() -> ValueMap {
     result.add_fn("print", |vm, args| {
         let result = match vm.get_args(args) {
             [Str(s)] => vm.stdout().write_line(s.as_str()),
-            [value] => vm.stdout().write_line(&value.to_string()),
+            [value] => {
+                let value = value.clone();
+                match vm.run_unary_op(crate::UnaryOp::Display, value)? {
+                    Str(s) => vm.stdout().write_line(s.as_str()),
+                    unexpected => {
+                        return unexpected_type_error_with_slice(
+                            "io.print",
+                            "string from @display",
+                            &[unexpected],
+                        )
+                    }
+                }
+            }
             [Str(format), format_args @ ..] => {
                 let format = format.clone();
                 let format_args = format_args.to_vec();